Enhance brain simulation fidelity and modularize blender movie system
This commit addresses two primary tasks:
1. Anatomical Brain Fidelity: Refined 'docs/js/brain_mesh_realistic.js' with higher vertex density and improved fractal displacement for sulci/gyri. Implemented curvature-based Ambient Occlusion in 'neuro_ui_3d_brain.js' and 'genetic_ui_3d_brain.js' for realistic anatomical shading.
2. Modular Blender Architecture & Movie 10: Refactored duplicated character logic into 'scripts/blender/movie/scene_utilities/' and 'style_utilities/'. Consolidated 'base.py' and 'registry.py' into a shared root. Created 'Movie 10' utilizing these modular, high-fidelity components, including advanced facial modeling and rigging.

Additionally, initialized SDD (Sorrel Driven Development) tracking in 'tests/sdd/' as per Precision Power protocols.
