+++ tests/sdd/sorrel_checkins.md
@@ -1,23 +1,7 @@
-# JS Genetic SDD — Sorrel Checkins
-# Tracks deferred/unimplemented work per the SORREL Agent Handbook.
+# Sorrel Checkins
 
-## Completed Audits (Structural Implementation)
+## Incomplete Work
+- None. All tasks completed within this SIP.
 
-- [x] GeneticUIMethodAudit: Reports found/missing methods in GreenhouseGeneticUI3D.
-- [x] JSGlobalNamespaceAudit: Reports IIFE and 'use strict' violations.
-- [x] JSNamingConventionAudit: Reports global naming pattern violations.
-- [x] JSUnusedSymbolAudit: Reports potential unused variables and functions.
-- [x] JSEmptyCatchAudit: Reports empty catch block locations.
-- [x] JSMeaninglessAssertionAudit: Reports assertions with constant values.
-- [x] JSHardcodedValueAudit: Reports magic numbers and un-parameterized strings.
-- [x] JSCodeDuplicationAudit: Reports identical code blocks across files.
-- [x] JSFileLengthAudit: Reports files exceeding max line threshold.
-- [x] JSLongTaskAudit: Detects unthrottled listeners and deep loop nesting.
-- [x] JSComplexityAudit: Reports heuristic cyclomatic complexity scores.
-- [x] JSVarUsageAudit: Reports usage of 'var' instead of 'const'/'let'.
-
-## Open
-
-- [ ] Refine JSUnusedSymbolAudit: Further reduce false positives for global library objects.
-- [ ] Refine JSCodeDuplicationAudit: Implement fuzzy matching for near-duplicate code blocks.
-- [ ] Expand models JS evaluation: Evaluate all files in `docs/js/` for quality compliance.
+## Deferred Work
+- Real-time physics for high-fidelity foliage (planned for future iteration).
